% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calculateDistance.R
\name{calculateGeographicalDistance}
\alias{calculateGeographicalDistance}
\title{Calculate geographical distance betwen two pointsin latitude and longitude in degrees}
\usage{
calculateGeographicalDistance(lon1, lon2, lat1, lat2)
}
\arguments{
\item{lon1}{longitude of point 1}

\item{lon2}{longitude o2 point 2}

\item{lat1}{latitude of point 1}

\item{lat2}{latitude of point 2}
}
\value{
numeric; distance in meters
}
\description{
Calculate geographical distance betwen two pointsin latitude and longitude in degrees
}
\examples{
calculateGeographicalDistance(57.60287, 57.62190, 11.66572, 11.67703)

}
