% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calculateDistance.R
\name{calculateGeographicalDistanceVectorized}
\alias{calculateGeographicalDistanceVectorized}
\title{Table with vessels trips data, columns needed: LON, LON_PREV, LAT, LAT_PREV}
\usage{
calculateGeographicalDistanceVectorized(vesselsTable)
}
\arguments{
\item{vesselsTable}{}
}
\value{
initial table with additional column distance (numeric)
}
\description{
Table with vessels trips data, columns needed: LON, LON_PREV, LAT, LAT_PREV
}
