% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation_depth.R
\name{compensation_depth}
\alias{compensation_depth}
\title{Import CTD data}
\usage{
compensation_depth(data)
}
\arguments{
\item{data}{The raw data read from CTD file.}
}
\value{
Compensation depth in unit \code{m}.
}
\description{
The purpose of this function is to calculate the compensation depth of a sampling site.
}
\details{
This function is based on \code{read_ctd()}, which is provided in this package. After getting raw data, this function can help you quickly get the compensation depth without going through complex data set.
}
\examples{
compensation_depth(data)

}
