% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-filter.R
\name{filterVesselsInput}
\alias{filterVesselsInput}
\title{Title}
\usage{
filterVesselsInput(id, vesselsTypes)
}
\arguments{
\item{vesselsTypes}{vector of characters, types of vessels (defined in global.R)}
}
\value{

}
\description{
Title
}
