% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-filter.R
\name{filterVesselsServer}
\alias{filterVesselsServer}
\title{Title}
\usage{
filterVesselsServer(id, vesselsTable, vesselsPerType)
}
\arguments{
\item{vesselsTable}{table with vesels defined in global.R}

\item{vesselsPerType}{list with vessels corresponding to types}
}
\value{
data.table; data corresponding to chosen vessel or NULL in case input is not ready
}
\description{
Title
}
