% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_processing.R
\name{get_data}
\alias{get_data}
\title{Read csv data into dataframe.}
\usage{
get_data(file = "ships_trim.csv")
}
\arguments{
\item{file}{File path.}
}
\description{
The file needs to be in the inst/extdata folder
}
\examples{
\dontrun{
df <- get_data()
df <- get_data(file = "file.csv")
}
}
