% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_processing.R
\name{get_gps_loc}
\alias{get_gps_loc}
\title{Get the longitude and latitude for given vessel ID.}
\usage{
get_gps_loc(df, vessel_id)
}
\arguments{
\item{df}{the dataframe containing the SHIPID column.}

\item{vessel_id}{SHIPID found in the dataframe.}
}
\description{
Looks up the GPS coordinate for the given vessel ID.
Finds the observation when it sailed the longest distance
between two consecutive observations
}
\examples{
\dontrun{
get_gps_loc(dataframe, vessel_id = "2764")
}
}
