% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_processing.R
\name{get_vessel_id}
\alias{get_vessel_id}
\title{Gets the list of unique vessel IDs.}
\usage{
get_vessel_id(df, ship_type = NULL)
}
\arguments{
\item{df}{the dataframe containing the SHIP_ID column.}

\item{ship_type}{the ship_type contained in the ship_type column of the df.}
}
\description{
Gets the list of unique vessels' ID.
The vessels' ID can be filtered based on vessel type.
}
\examples{
\dontrun{
get_vessel_id(dataframe)
get_vessel_id(dataframe, ship_type == "Cargo")
}
}
