% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_processing.R
\name{get_vessel_type}
\alias{get_vessel_type}
\title{List of unique vessel types.}
\usage{
get_vessel_type(df)
}
\arguments{
\item{df}{the dataframe containing the ship_type column.}
}
\description{
Gets the list of unique vessels' types.
It looks in the ship_type column of the dataframe and
returned the list of unique types.
}
\examples{
\dontrun{
get_vessel_type(dataframe)
}
}
