% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{gps}
\alias{gps}
\title{Import CTD data}
\usage{
gps(data)
}
\arguments{
\item{data}{The raw data read from CTD file.}
}
\value{
A list containing both longitude and latitude of the sampling site.
}
\description{
The purpose of this function is to get GPS information from CTD data.
}
\details{
This function is based on \code{read_ctd()}, which is provided in this package. After getting raw data, this function can help you quickly get the GPS data without going through complex data set.
}
\examples{
gps(data)

}
