% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marine-package.r
\docType{package}
\name{marine}
\alias{marine}
\alias{marine-package}
\title{CTD Data Quick Analysis}
\description{
Dealing with CTD data set is difficult. First, the .cnv file is a unique format and no applications can read it. Second, the data structure is quite complex. Thus, I designed this package to help researchers get a quick analysis of CTD data.
}
\details{
For detailed instruction of how to use this package, please see README or vignettes.
}
\note{
This package (Version 0.1.0) is designed for STAT547M homework. Updates will be made later for research purposes. Please contact me with your research needs so that I can add more functions into this package.
}
\author{
Yuanji Sun
}
