% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ctd.R
\name{raw_data}
\alias{raw_data}
\title{Import CTD data}
\usage{
raw_data(file)
}
\arguments{
\item{file}{The file to be read. It must be a .cnv file.}
}
\value{
A R object containing all raw data in .cnv file. This data is very complex, which needs further process.
}
\description{
The purpose of this function is to import CTD data as an variable in R (raw_data).
}
\details{
Since .cnv file is quite unique and the data structure is extremely complex, this function is based on \code{read.oce()} in the package \code{oce}.
}
\examples{
raw_data("sample_CTD.cnv")

}
