% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calculateDistance.R
\name{selectRowWithLargestDistance}
\alias{selectRowWithLargestDistance}
\title{Table returned by calculateGeographicalDistanceVectorized function, columns needed: SHIP_ID, DATETIME, distance}
\usage{
selectRowWithLargestDistance(vesselsTable)
}
\arguments{
\item{vesselsTable}{}
}
\value{
table with one row corresponding to the longest distance found
}
\description{
Table returned by calculateGeographicalDistanceVectorized function, columns needed: SHIP_ID, DATETIME, distance
}
