% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Sample raw data from .cnv file.}
\format{The format is a complex dataframe, which is nested. This is only a test data set used in the testthat.}
\source{
Documenting data.
}
\usage{
test_data
}
\description{
Sample raw data from .cnv file.
}
\keyword{datasets}
