% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data}
\alias{data}
\title{Sample raw data from .cnv file.}
\format{The format is a complex dataframe, which is nested.}
\source{
Documenting data.
}
\usage{
data(..., list = character(), package = NULL, lib.loc = NULL,
  verbose = getOption("verbose"), envir = .GlobalEnv)
}
\description{
Sample raw data from .cnv file.
}
