% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-filterVassels.R
\name{filterVesselsServer}
\alias{filterVesselsServer}
\title{Title}
\usage{
filterVesselsServer(id)
}
\arguments{
\item{id}{}
}
\value{
data.table; row of data corresponding to chosen vessel
}
\description{
Title
}
