% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-filter.R
\name{filterVesselsServer}
\alias{filterVesselsServer}
\title{Title}
\usage{
filterVesselsServer(id)
}
\arguments{
\item{id}{}
}
\value{
data.table; data corresponding to chosen vessel or NULL in case input is not ready
}
\description{
Title
}
